Split participant_status.py into append.py + lifecycle.py submodules
- Move 5 append-sequence nodes (SkipIfIdempotentNode, LoadParticipantNode,
  CheckStatusNotAlreadyAppendedNode, ResolveAndPersistStatusObjectNode,
  ValidateRMTransitionNode, AppendStatusAndSaveParticipantNode) from
  participant_status.py into nodes/append.py (DEMOMA-07-003 step 2)
- Move 3 lifecycle nodes (_PublicDisclosureSkipConditionNode,
  PublicDisclosureBranchNode, AutoCloseBranchNode) into nodes/lifecycle.py
  (DEMOMA-07-003 steps 4-5)
- Remove inline SkipIfIdempotentNode closure from append_participant_status_tree.py
  (BTND-07-003: leaf nodes must not be defined inside tree factory functions)
- SkipIfIdempotentNode now accepts explicit constructor args instead of
  closing over enclosing scope
- Update nodes/__init__.py to import from append + lifecycle (not participant_status)
- Split test_participant_status.py into test_append.py + test_lifecycle.py;
  add dedicated SkipIfIdempotentNode tests (not/already-appended cases)
- participant_status.py deleted; all 3191 tests pass, all linters clean

Fixes CS-18-002 (append.py=407 lines, lifecycle.py=172 lines, both ≤500)
Fixes BTND-07-003 (SkipIfIdempotentNode moved out of tree factory file)
