+++ vultron/core/behaviors/status/nodes/append.py
@@ -13,11 +13,11 @@
 #  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
 #  U.S. Patent and Trademark Office by Carnegie Mellon University
 
-"""Participant status workflow nodes for AddParticipantStatusToParticipant.
+"""Append-participant-status leaf nodes for DEMOMA-07-003 step 2.
 
-Contains the load, idempotency check, status resolution, RM transition
-validation, append, public disclosure branch, and auto-close nodes that
-implement steps 2–5 of DEMOMA-07-003.
+Contains the five leaf nodes that implement the append sequence:
+load participant, check idempotency, resolve status object, validate RM
+transition, and append + save.
 """
 
 import logging
@@ -26,23 +26,67 @@ from typing import Any, cast
 import py_trees
 from py_trees.common import Status
 
-from vultron.core.behaviors.embargo.nodes import TerminateEmbargoNode
 from vultron.core.behaviors.helpers import DataLayerAction, DataLayerCondition
-from vultron.core.behaviors.status.nodes.broadcast import _find_case_manager_id
 from vultron.core.models.protocols import (
     PersistableModel,
-    is_case_model,
     is_participant_model,
 )
 from vultron.core.states.rm import (
-    RM,
     is_monotonic_rm_forward,
     is_valid_rm_transition,
 )
-from vultron.core.states.roles import CVDRole
 from vultron.core.use_cases._helpers import _as_id
 
 logger = logging.getLogger(__name__)
+
+
+class SkipIfIdempotentNode(py_trees.behaviour.Behaviour):
+    """Idempotency guard for the append-participant-status Selector.
+
+    Returns SUCCESS when *status_id* is already present in the participant's
+    status list — causing the parent Selector to short-circuit and skip the
+    append subtree. Returns FAILURE when the status is not yet appended,
+    allowing the parent Selector to continue to the append subtree.
+
+    This is the inverse of :class:`CheckStatusNotAlreadyAppendedNode`: that
+    node is used to halt a Sequence on duplicate; this node is used to skip
+    an append Selector on duplicate.
+
+    Per DEMOMA-07-003 step 2 idempotency requirement.
+    """
+
+    def __init__(
+        self,
+        status_id: str,
+        participant_id: str,
+        name: str | None = None,
+    ):
+        super().__init__(name=name or self.__class__.__name__)
+        self.status_id = status_id
+        self.participant_id = participant_id
+
+    def setup(self, **kwargs: Any) -> None:
+        self.blackboard = py_trees.blackboard.Client(name=self.name)
+        self.blackboard.register_key(
+            key="append_status_participant",
+            access=py_trees.common.Access.READ,
+        )
+
+    def update(self) -> Status:
+        participant = self.blackboard.get("append_status_participant")
+        if participant is None:
+            return Status.FAILURE
+
+        existing_ids = [_as_id(s) for s in participant.participant_statuses]
+        if self.status_id in existing_ids:
+            logging.getLogger(self.__class__.__module__).info(
+                "SkipIfIdempotentNode: status '%s' already on participant"
+                " '%s' — idempotent, skipping (SUCCESS)",
+                self.status_id,
+                self.participant_id,
+            )
+            return Status.SUCCESS
+        return Status.FAILURE
 
 
 class LoadParticipantNode(DataLayerAction):
@@ -361,200 +405,5 @@ class AppendStatusAndSaveParticipantNode(DataLayerAction):
             " participant '%s' (DEMOMA-07-003 step 2)",
             self.status_id,
             self.participant_id,
-        )
-        return Status.SUCCESS
-
-
-class _PublicDisclosureSkipConditionNode(DataLayerCondition):
-    """Inner guard for :class:`PublicDisclosureBranchNode`.
-
-    Returns SUCCESS (skip teardown) when:
-    - The new status is NOT public-aware (CS.P not set), OR
-    - DataLayer or case_id is unavailable, OR
-    - The sender is not a known case participant, OR
-    - The sender does NOT hold the CASE_OWNER role.
-
-    Returns FAILURE (proceed to teardown) when the sender IS a CASE_OWNER
-    who has sent a public-aware status update.
-    """
-
-    def __init__(
-        self,
-        status_obj: PersistableModel | None,
-        sender_actor_id: str,
-        case_id: str | None,
-        name: str | None = None,
-    ):
-        super().__init__(name=name or self.__class__.__name__)
-        self.status_obj = status_obj
-        self.sender_actor_id = sender_actor_id
-        self.case_id = case_id
-
-    def _public_aware(self) -> bool:
-        """Return True if the status signals public awareness (CS.P is set)."""
-        from vultron.core.states.cs import CS_pxa
-
-        case_status = getattr(self.status_obj, "case_status", None)
-        pxa_state = getattr(case_status, "pxa_state", None)
-        if pxa_state is None:
-            return False
-        try:
-            return pxa_state in (
-                CS_pxa.Pxa,
-                CS_pxa.PxA,
-                CS_pxa.PXa,
-                CS_pxa.PXA,
-            )
-        except Exception:
-            return False
-
-    def update(self) -> Status:
-        if not self._public_aware():
-            return Status.SUCCESS
-
-        if self.datalayer is None or not self.case_id:
-            return Status.SUCCESS
-
-        case = self.datalayer.read(self.case_id)
-        if not is_case_model(case):
-            return Status.SUCCESS
-
-        sender_participant_id = case.actor_participant_index.get(
-            self.sender_actor_id
-        )
-        if sender_participant_id is None:
-            return Status.SUCCESS
-
-        sender_participant = self.datalayer.read(sender_participant_id)
-        roles = getattr(sender_participant, "case_roles", [])
-        if CVDRole.CASE_OWNER not in roles:
-            return Status.SUCCESS
-
-        # Condition met: sender is CASE_OWNER reporting public awareness.
-        return Status.FAILURE
-
-
-class PublicDisclosureBranchNode(py_trees.composites.Selector):
-    """Step 4: Trigger embargo teardown if public disclosure is detected.
-
-    Condition: the new ParticipantStatus has CS.P (public-aware) set AND
-    the sender holds the CASE_OWNER role.
-
-    When the condition is met, delegates to :class:`TerminateEmbargoNode`.
-    Skips silently if conditions are not met or trigger_activity_factory
-    is unavailable.
-
-    Always returns SUCCESS (failure to initiate teardown is not fatal to
-    the parent sequence).
-
-    Implemented as a ``py_trees.composites.Selector`` (memory=False):
-    - Child 1 ``_PublicDisclosureSkipConditionNode``: SUCCESS → skip teardown.
-    - Child 2 ``TerminateEmbargoNode``: always SUCCESS → teardown attempted.
-
-    Per DEMOMA-07-003 step 4.
-    """
-
-    def __init__(
-        self,
-        status_obj: PersistableModel | None,
-        sender_actor_id: str,
-        case_id: str | None,
-        name: str | None = None,
-    ):
-        super().__init__(name=name or self.__class__.__name__, memory=False)
-        self.add_children(
-            [
-                _PublicDisclosureSkipConditionNode(
-                    status_obj=status_obj,
-                    sender_actor_id=sender_actor_id,
-                    case_id=case_id,
-                    name="SkipCondition",
-                ),
-                TerminateEmbargoNode(
-                    case_id=case_id,
-                    name="TerminateEmbargo",
-                ),
-            ]
-        )
-
-
-class AutoCloseBranchNode(DataLayerAction):
-    """Step 5: Log case auto-close when all participants are RM.CLOSED.
-
-    Checks whether every CVD participant in the case has ``RM.CLOSED``
-    as their most recent status.  The CASE_MANAGER (Case Actor) is a
-    coordinator role and is excluded from this check.
-
-    Actual case closure is **not** persisted here — this is log-only
-    behaviour for the prototype demo.
-
-    Always returns SUCCESS.
-
-    Per DEMOMA-07-003 step 5.
-    """
-
-    def __init__(
-        self,
-        case_id: str | None,
-        name: str | None = None,
-    ):
-        super().__init__(name=name or self.__class__.__name__)
-        self.case_id = case_id
-
-    def _all_participants_closed(self, case: Any) -> bool:
-        """Return True iff every CVD participant has RM.CLOSED."""
-        if self.datalayer is None:
-            return False
-
-        for p_id in case.actor_participant_index.values():
-            p = self.datalayer.read(p_id)
-            if p is None:
-                return False
-            roles = getattr(p, "case_roles", [])
-            if CVDRole.CASE_MANAGER in roles:
-                continue
-            statuses = getattr(p, "participant_statuses", [])
-            if not statuses:
-                return False
-            latest_ref = statuses[-1]
-            if isinstance(latest_ref, str):
-                ref_id = _as_id(latest_ref)
-                if ref_id is None:
-                    return False
-                latest = self.datalayer.read(ref_id)
-            else:
-                latest = latest_ref
-            if latest is None:
-                return False
-            rm_state = getattr(latest, "rm_state", None)
-            if rm_state is None or rm_state != RM.CLOSED:
-                return False
-        return True
-
-    def update(self) -> Status:
-        if self.datalayer is None or not self.case_id:
-            return Status.SUCCESS
-
-        case = self.datalayer.read(self.case_id)
-        if not is_case_model(case):
-            return Status.SUCCESS
-
-        if not self._all_participants_closed(case):
-            return Status.SUCCESS
-
-        case_manager_id = _find_case_manager_id(self.datalayer, case)
-        if case_manager_id is None:
-            self.logger.warning(
-                "AutoCloseBranch: no Case Manager found"
-                " — cannot auto-close case '%s'",
-                self.case_id,
-            )
-            return Status.SUCCESS
-
-        self.logger.info(
-            "AutoCloseBranch: Case Manager '%s' auto-closing case '%s'"
-            " — all participants CLOSED (DEMOMA-07-003 step 5)",
-            case_manager_id,
-            self.case_id,
         )
         return Status.SUCCESS
